refactor: ProgressStatCard to own file
