+++ screens/userStats/UserStatsScreen.js
@@ -4,7 +4,8 @@ import {
     View,
     ScrollView,
     Pressable,
-    ActivityIndicator
+    ActivityIndicator,
+    Dimensions
 } from 'react-native';
 import AsyncStorage from '@react-native-community/async-storage';
 import * as actions from '../../actions';
@@ -16,29 +17,10 @@ import {
     AnimatedCircle,
     Header,
     Colors,
-    StatsGrid,
-    Caption
+    StatsGrid
 } from '../components';
-
-const ProgressStatCard = ({ value, title, tagline, color, style }) => {
-    return (
-        <View
-            style={[
-                style,
-                {
-                    // borderTopColor: color,
-                    // borderTopWidth: 4,
-                    borderRadius: 8,
-                    paddingLeft: 10,
-                    paddingVertical: 10
-                }
-            ]}>
-            <Title style={{ color }}>{value}</Title>
-            <Body style={{ color }}>{title}</Body>
-            <Caption>{tagline}</Caption>
-        </View>
-    );
-};
+import { ProgressStatCard } from './_components';
+const { width } = Dimensions.get('window');
 
 class UserStatsScreen extends Component {
     constructor(props) {
@@ -197,7 +179,7 @@ class UserStatsScreen extends Component {
                                     value={user.level}
                                     delay={500}
                                     duration={5000}
-                                    radius={70}
+                                    radius={(width - 40) / 4 - 16}
                                     // tagline="Level"
                                 />
                             </View>
@@ -207,11 +189,11 @@ class UserStatsScreen extends Component {
                                     isValueDisplayed={false}
                                     strokeWidth={10}
                                     percentage={user.targetPercentage + 10}
-                                    color={`#f7ca18`}
+                                    color="#f7ca18"
                                     value={user.level}
                                     delay={500}
                                     duration={5000}
-                                    radius={90}
+                                    radius={(width - 40) / 4}
                                     tagline="Level"
                                 />
                             </View>
