+++ screens/userStats/_components/index.js
@@ -1,2 +1,3 @@
 export { default as StatsCard } from './StatsCard';
 export { default as RewardsList } from './RewardsList';
+export { default as ProgressStatCard } from './ProgressStatCard';
